Xray dev (#84)
* finish proxy; move to certbot

* remove test.sh

* minor changes

* colorEcho -> echo

* fix spelling

* change proxy addr; imporve logging

* fix certbot; fix typo

* add rm compatibility for older veriosn

* ready to merge

+++ tools/rm_xwall.sh
@@ -51,21 +51,41 @@ uninstall() {
 }
 
 # remove v2ray
-colorEcho ${BLUE} "Stopping v2ray service."
-${sudoCmd} systemctl stop v2ray
-${sudoCmd} systemctl disable v2ray
-uninstall "/etc/systemd/system/v2ray.service"
-colorEcho ${BLUE} "Removing v2ray binaries."
-uninstall "/usr/local/bin/v2ray"
-uninstall "/usr/local/bin/v2ctl"
-colorEcho ${BLUE} "Removing v2ray files."
-uninstall "/usr/local/etc/v2ray"
-uninstall"/usr/local/share/v2ray"
-uninstall "/var/log/v2ray"
-colorEcho ${BLUE} "Removing v2ray crontab."
-${sudoCmd} crontab -l | grep -v 'v2ray/geoip.dat' | ${sudoCmd} crontab -
-${sudoCmd} crontab -l | grep -v 'v2ray/geosite.dat' | ${sudoCmd} crontab -
-colorEcho ${GREEN} "Removed v2ray successfully."
+if [ -f "/usr/local/bin/v2ray" ]; then
+  colorEcho ${BLUE} "Stopping v2ray service."
+  ${sudoCmd} systemctl stop v2ray
+  ${sudoCmd} systemctl disable v2ray
+  uninstall "/etc/systemd/system/v2ray.service"
+  colorEcho ${BLUE} "Removing v2ray binaries."
+  uninstall "/usr/local/bin/v2ray"
+  uninstall "/usr/local/bin/v2ctl"
+  colorEcho ${BLUE} "Removing xray files."
+  uninstall "/usr/local/etc/v2ray"
+  uninstall"/usr/local/share/v2ray"
+  uninstall "/var/log/v2ray"
+  colorEcho ${BLUE} "Removing xray crontab."
+  ${sudoCmd} crontab -l | grep -v 'v2ray/geoip.dat' | ${sudoCmd} crontab -
+  ${sudoCmd} crontab -l | grep -v 'v2ray/geosite.dat' | ${sudoCmd} crontab -
+  colorEcho ${GREEN} "Removed xray successfully."
+fi
+
+# remove xray
+if [ -f "/usr/local/bin/xray" ]; then
+  colorEcho ${BLUE} "Stopping xray service."
+  ${sudoCmd} systemctl stop xray
+  ${sudoCmd} systemctl disable xray
+  uninstall "/etc/systemd/system/xray.service"
+  colorEcho ${BLUE} "Removing xray binaries."
+  uninstall "/usr/local/bin/xray"
+  colorEcho ${BLUE} "Removing xray files."
+  uninstall "/usr/local/etc/xray"
+  uninstall"/usr/local/share/xray"
+  uninstall "/var/log/xray"
+  colorEcho ${BLUE} "Removing xray crontab."
+  ${sudoCmd} crontab -l | grep -v 'xray/geoip.dat' | ${sudoCmd} crontab -
+  ${sudoCmd} crontab -l | grep -v 'xray/geosite.dat' | ${sudoCmd} crontab -
+  colorEcho ${GREEN} "Removed xray successfully."
+fi
 
 # remove trojan-go
 if [ -f "/usr/bin/trojan-go" ]; then
@@ -81,6 +101,10 @@ if [ -f "/usr/bin/trojan-go" ]; then
   uninstall /etc/trojan-go
   colorEcho ${GREEN} "Removed trojan-go successfully."
 fi
+
+colorEcho ${BLUE} "Removing dummy site."
+${sudoCmd} $(which rm) -rf /var/www/acme
+${sudoCmd} $(which rm) -rf /var/www/html/*
 
 # remove naive
 if [ -f "/usr/local/bin/naive" ]; then
@@ -110,28 +134,19 @@ if [ -f "/usr/local/bin/caddy" ]; then
   colorEcho ${GREEN} "Removed caddy successfully."
 fi
 
-# remove nginx
-if [ -f "/etc/systemd/system/nginx.service" ]; then
-  colorEcho ${BLUE} "Shutting down nginx service."
-  ${sudoCmd} systemctl stop nginx
-  ${sudoCmd} systemctl disable nginx
-  uninstall /etc/systemd/system/nginx.service
-  colorEcho ${BLUE} "Removing nginx"
-  ${sudoCmd} ${systemPackage} remove nginx -y
-  ${sudoCmd} ${systemPackage} remove nginx-full -y
-  if [ -d "/usr/local/nginx" ]; then
-    uninstall /usr/local/nginx
-  fi
-  colorEcho ${GREEN} "Removed nginx successfully."
+# remove acme.sh
+if [ -d "/root/.acme.sh" ] || [ -d "~/.acme.sh" ]; then
+  colorEcho ${BLUE} "Removing acme.sh"
+  ${sudoCmd} bash /root/.acme.sh/acme.sh --uninstall
+  uninstall /root/.acme.sh
+  colorEcho ${GREEN} "Removed acme.sh successfully."
 fi
 
-colorEcho ${BLUE} "Removing dummy site."
-${sudoCmd} $(which rm) -rf /var/www/html
-
-# remove acme.sh
-colorEcho ${BLUE} "Removing acme.sh"
-${sudoCmd} bash /root/.acme.sh/acme.sh --uninstall
-uninstall /root/.acme.sh
+# remove cerbot
+colorEcho ${BLUE} "Removing certbot"
+${sudoCmd} ${systemPackage} remove certbot -y
+uninstall /etc/letsencrypt
+${sudoCmd} crontab -l | grep -v 'certbot' | ${sudoCmd} crontab -
 colorEcho ${GREEN} "Removed acme.sh successfully."
 
 colorEcho ${BLUE} "卸载完成"
